Rename frame_start and frame_end to frame_first and frame_last
frame_start was also used as the method to run the FRAME GO
command. The first name change resolves the conflict and the second
change ensures that the styles of the two names match (start and end
becomes first and last, not first and end).

Also: bugfix, norm_time was referred to without the class name twice.